Home activity and Add Training UI design
Adding, editing and removing users is working well on basic
functionality without any additional features at the moment. All files
are stored locally.
Home activity consist of basic information and is very limited at this
point.
Started to work on the UI design for adding the training.
Whole application will be refactored and multiple things will require
entirely different approach as this is no longer school project, but
personal one and more time can be invested in it with the design of
better working features.